printf has been added

+++ include/chess.h
@@ -1,7 +1,6 @@
 #ifndef CHESS_H
 #define CHESS_H
 
-#include "config.h"
 
 class Piece
 {
@@ -11,7 +10,7 @@ public:
     Piece(){}
     Piece(int x, int y);
     ~Piece(){}
-    virtual void posMov(std::string *, int *, int *) = 0;
+    virtual void posMov(std::string *, float *, float *) = 0;
     virtual void setxPos(int) = 0;
     virtual void setyPos(int) = 0;
 };
@@ -24,7 +23,7 @@ public:
     KnightPiece();
     ~KnightPiece();
 
-    void posMov(std::string *, int *, int *);
+    void posMov(std::string *, float *, float *);
     void setxPos(int);
     void setyPos(int);
     
@@ -36,7 +35,7 @@ public:
     BishopPiece();
     ~BishopPiece();
 
-    void posMov(std::string *, int *, int *);
+    void posMov(std::string *, float *, float *);
     void setxPos(int);
     void setyPos(int);
 };
